# cmd/filmlibrary/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Film Library Application",
        "title": "Film Library API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
        "/actors": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "List all actors with optional query parameters 'limit' and 'offset'",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actors"
                ],
                "summary": "List all actors with optional query parameters 'limit' and 'offset'",
                "operationId": "list-actors",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "An optional query parameter 'limit' that limits total number of returned actors. By default 'limit' = 100",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "An optional query parameter 'offset' that indicates how many records should be skipped while listing actors. By default 'offset' = 0",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_app_entrypoint_http.listActorsResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Create a new actor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actors"
                ],
                "summary": "Create a new actor",
                "operationId": "create-actor",
                "parameters": [
                    {
                        "description": "Actor object that needs to be created",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_entrypoint_http.createActorRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_app_entrypoint_http.createActorResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    }
                }
            }
        },
...
        "/films": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "List all films with the possibility of sorting via 'sort' parameter by 'title' and/or 'rating' and/or 'release-date' and/or with 'limit' and/or 'offset' parameters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "films"
                ],
                "summary": "List all films with optional 'sort', 'limit', 'offset' query parameters",
                "operationId": "list-films",
                "parameters": [
                    {
                        "type": "string",
                        "description": "An optional query parameter 'sort' that indicates how films should be sorted. By default 'sort' = 'rating:desc'. Expected as `title:asc,release-date:desc,rating:desc` in any order of necessary parameters",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "An optional query parameter 'limit' that limits total number of returned films. By default 'limit' = 100",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "An optional query parameter 'offset' that indicates how many records should be skipped while listing films. By default 'offset' = 0",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_app_entrypoint_http.listFilmsResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Create a new film",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "films"
                ],
                "summary": "Create a new film",
                "operationId": "create-film",
                "parameters": [
                    {
                        "description": "Film object that needs to be created",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_entrypoint_http.createFilmRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_app_entrypoint_http.createFilmResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response": {
            "type": "object",
            "properties": {
                "payload": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "internal_app_entrypoint_http.actor": {
            "type": "object",
            "properties": {
                "birthdate": {
                    "type": "string"
                },
                "films": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_app_entrypoint_http.actorFilm"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "sex": {
                    "type": "integer"
                }
            }
        },
        "internal_app_entrypoint_http.actorFilm": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "rating": {
                    "type": "integer"
                },
                "release_date": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "internal_app_entrypoint_http.createActorRequestBody": {
            "type": "object",
            "required": [
                "birthdate",
                "name",
                "sex"
            ],
            "properties": {
                "birthdate": {
                    "type": "string",
                    "example": "2006-01-02"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100
                },
                "sex": {
                    "type": "integer",
                    "enum": [
                        0,
                        1,
                        2,
                        9
                    ]
                }
            }
        },
        "internal_app_entrypoint_http.createActorResponseBody": {
            "type": "object",
            "properties": {
                "birthdate": {
                    "type": "string"
                },
                "films": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_app_entrypoint_http.actorFilm"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "sex": {
                    "type": "integer"
                }
            }
        },
        "internal_app_entrypoint_http.createFilmRequestBody": {
            "type": "object",
            "required": [
                "actor_ids",
                "rating",
                "release_date",
                "title"
            ],
            "properties": {
                "actor_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        1,
                        2,
                        3
                    ]
                },
                "description": {
                    "type": "string",
                    "maxLength": 1000
                },
                "rating": {
                    "type": "integer",
                    "maximum": 10,
                    "minimum": 0
                },
                "release_date": {
                    "type": "string",
                    "example": "2006-01-02"
                },
                "title": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 1
                }
            }
        },
        "internal_app_entrypoint_http.createFilmResponseBody": {
            "type": "object",
            "properties": {
                "actors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_app_entrypoint_http.filmActor"
                    }
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "rating": {
                    "type": "integer"
                },
                "release_date": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "internal_app_entrypoint_http.film": {
            "type": "object",
            "properties": {
                "actors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_app_entrypoint_http.filmActor"
                    }
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "rating": {
                    "type": "integer"
                },
                "release_date": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "internal_app_entrypoint_http.filmActor": {
            "type": "object",
            "properties": {
                "birthdate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "sex": {
                    "type": "integer"
                }
            }
        },
        "internal_app_entrypoint_http.listActorsResponseBody": {
            "type": "object",
            "properties": {
                "actors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_app_entrypoint_http.actor"
                    }
                }
            }
        },
        "internal_app_entrypoint_http.listFilmsResponseBody": {
            "type": "object",
            "properties": {
                "films": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_app_entrypoint_http.film"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}